refactor after code review comments, part 1
