+++ contracts/sg-eth-airdrop/rustfmt.toml
@@ -12,4 +12,3 @@ tab_spaces = 4
 #struct_field_align_threshold = 20
 #struct_lit_single_line = true
 #report_todo = "Always"
-
