--- contracts/sg-eth-airdrop/src/helpers/constants.rs
@@ -1,11 +1,0 @@
-pub const GENERIC_WHITELIST_LABEL: &str = "Generic Whitelist for Airdrop";
-pub const NATIVE_DENOM: &str = "ustars";
-pub const CONTRACT_NAME: &str = "crates.io:sg-eth-airdrop";
-pub const CONTRACT_VERSION: &str = env!("CARGO_PKG_VERSION");
-pub const INIT_WHITELIST_REPLY_ID: u64 = 1;
-pub const COLLECTION_WHITELIST: &str = "contract1";
-
-pub const INSTANTIATION_FEE: u128 = 100_000_000; // 100 STARS
-
-pub const MIN_AIRDROP: u128 = 10_000_000; // 10 STARS
-pub const MAX_AIRDROP: u128 = 100_000_000_000_000; // 100 million STARS
