--- contracts/sg-eth-airdrop/src/helpers/lib.rs
@@ -1,23 +1,0 @@
-mod build_msg;
-pub use build_msg::{
-    build_add_member_minter_msg, build_bank_message, build_config_msg,
-    build_messages_for_claim_and_whitelist_add, build_whitelist_instantiate_msg,
-};
-
-mod validation;
-pub use validation::{
-    check_instantiate_funds, run_validations_for_claim, validate_airdrop_amount, validate_eth_sig,
-    validate_mints_remaining, validate_plaintext_msg,
-};
-
-mod funds;
-pub use funds::check_funds_and_fair_burn;
-
-mod constants;
-pub use constants::{
-    COLLECTION_WHITELIST, CONTRACT_NAME, CONTRACT_VERSION, GENERIC_WHITELIST_LABEL,
-    INIT_WHITELIST_REPLY_ID, INSTANTIATION_FEE, MAX_AIRDROP, MIN_AIRDROP, NATIVE_DENOM,
-};
-
-mod computation;
-pub use computation::{compute_plaintext_msg, compute_valid_eth_sig};
