+++ contracts/sg-eth-airdrop/src/lib.rs
@@ -1,16 +1,12 @@
 pub mod contract;
 mod error;
 pub mod msg;
-mod query;
+pub mod query;
 pub mod reply;
 pub mod state;
 pub use crate::error::ContractError;
+mod claim_airdrop;
 
-#[path = "./ethereum/lib.rs"]
-mod ethereum;
-
-#[path = "./helpers/lib.rs"]
-pub mod helpers;
 #[cfg(test)]
 #[path = "./testing/lib.rs"]
 pub mod tests_folder;
